Amlet.jl Documentation

Amlet.jl Documentation

Types

Functions

Objective

Gradient

Hessian

Hessian approxination : BHHH

Index

Types

Amlet.LogitUtility — Type

LogitUtility{L <: isLinear}

Fields

u::Function : utility function for Logit model

grad::Function : gradient

x::Any

beta::AbstractVector

i::Int

H::Function : Hessian matrix

x::Any

beta::AbstractVector

i::Int

Functions

Objective

This is the computation of the objective function

\[-\dfrac{1}{N} \sum_{i=1}^{N} \sum_{j=1}^{n_{alt}} y_{ij}\log\big( \mathbb{P}(j \vert \beta , X) \big) \]

Sofia.F — Function

F(x, mo; sample , update)

Function to compute the averaged negative log-likelihood of the Logit model over a sampled population.

Arguments

x::Vector{T}` is the parameter vector

mo::LogitModel{UPD, D} is the Logit modelwhere data is stored

sample = 1:length(mo.data) is the sampled population used for computation, by default full.

update::Bool = false is for use of storage Engine.

Sofia.Fs — Function

Fs(x, mo; sample)

Computation of the negative log-likelihoods for all individuals in the selected population.

Returns an array comtaining all values.

Arguments

x::Vector{T}` is the parameter vector

mo::LogitModel{UPD, D} is the Logit modelwhere data is stored

sample = 1:length(mo.data) is the sampled population used for computation, by default full.

Sofia.Fs! — Function

Fs!(x, mo, ac; sample , inplace)

Computation of the negative log-likelihoods for all individuals in the selected population.

Modifies the ac array in place.

Arguments

x::Vector{T}` is the parameter vector

mo::LogitModel{UPD, D} is the Logit modelwhere data is stored

ac::Array is the accumulator vector to modify in place

sample = 1:length(mo.data) is the sampled population used for computation, by default full

inplace::Bool = false

Gradient

Missing docstring.

Missing docstring for grad. Check Documenter's build log for details.

Missing docstring.

Missing docstring for grad!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for grads. Check Documenter's build log for details.

Missing docstring.

Missing docstring for grads!. Check Documenter's build log for details.

Hessian

Missing docstring.

Missing docstring for H. Check Documenter's build log for details.

Missing docstring.

Missing docstring for H!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Hdotv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Hdotv!. Check Documenter's build log for details.

Hessian approxination : BHHH

Missing docstring.

Missing docstring for BHHH. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BHHH!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BHHHdotv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BHHHdotv!. Check Documenter's build log for details.

Index

Amlet.LogitUtility

Sofia.F

Sofia.Fs

Sofia.Fs!